move bid method to bid file
